Move restrictions and data value types to dedicated files
